+++ src/test/java/org/cicirello/search/evo/GenerationalEATests.java
@@ -33,7 +33,8 @@ import org.cicirello.search.operators.MutationOperator;
 import org.cicirello.search.operators.CrossoverOperator;
 
 /**
- * JUnit 4 test cases for GenerationalEvolutionaryAlgorithm 
+ * JUnit test cases for GenerationalEvolutionaryAlgorithm,
+ * GenerationalNANDOperatorsEvolutionaryAlgorithm,  
  * and GenerationalMutationOnlyEvolutionaryAlgorithm.
  */
 public class GenerationalEATests {
@@ -122,13 +123,12 @@ public class GenerationalEATests {
 		);
 		thrownNull = assertThrows( 
 			NullPointerException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 				5, 
 				null, 
 				0.5, 
 				new CountCrossoverCalls(),
 				0.5,
-				false,
 				new TestInitializer(), 
 				new TestFitnessDouble(), 
 				new TestSelectionOp(), 
@@ -137,13 +137,12 @@ public class GenerationalEATests {
 		);
 		thrownNull = assertThrows( 
 			NullPointerException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 				5, 
 				null, 
 				0.5,
 				new CountCrossoverCalls(),
 				0.5,
-				false,
 				new TestInitializer(), 
 				new TestFitnessInteger(), 
 				new TestSelectionOp(), 
@@ -152,13 +151,12 @@ public class GenerationalEATests {
 		);
 		thrownNull = assertThrows( 
 			NullPointerException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 				5, 
 				new CountMutationCalls(), 
 				0.5, 
 				null,
 				0.5,
-				false,
 				new TestInitializer(), 
 				new TestFitnessDouble(), 
 				new TestSelectionOp(), 
@@ -167,13 +165,12 @@ public class GenerationalEATests {
 		);
 		thrownNull = assertThrows( 
 			NullPointerException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 				5, 
 				new CountMutationCalls(), 
 				0.5,
 				null,
 				0.5,
-				false,
 				new TestInitializer(), 
 				new TestFitnessInteger(), 
 				new TestSelectionOp(), 
@@ -262,13 +259,68 @@ public class GenerationalEATests {
 		);
 		thrownIllegal = assertThrows( 
 			IllegalArgumentException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 				5, 
 				new CountMutationCalls(), 
 				0.5, 
 				new CountCrossoverCalls(),
 				0.5+1E-10,
-				true,
+				new TestInitializer(), 
+				new TestFitnessInteger(), 
+				new TestSelectionOp(), 
+				new ProgressTracker<TestObject>()
+			)
+		);
+		thrownIllegal = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
+				5, 
+				new CountMutationCalls(), 
+				1E-10, 
+				new CountCrossoverCalls(),
+				-1E-10,
+				new TestInitializer(), 
+				new TestFitnessDouble(), 
+				new TestSelectionOp(), 
+				new ProgressTracker<TestObject>()
+			)
+		);
+		thrownIllegal = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
+				5, 
+				new CountMutationCalls(), 
+				1E-10, 
+				new CountCrossoverCalls(),
+				-1E-10,
+				new TestInitializer(), 
+				new TestFitnessInteger(), 
+				new TestSelectionOp(), 
+				new ProgressTracker<TestObject>()
+			)
+		);
+		thrownIllegal = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
+				5, 
+				new CountMutationCalls(), 
+				-1E-10, 
+				new CountCrossoverCalls(),
+				1E-10,
+				new TestInitializer(), 
+				new TestFitnessDouble(), 
+				new TestSelectionOp(), 
+				new ProgressTracker<TestObject>()
+			)
+		);
+		thrownIllegal = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
+				5, 
+				new CountMutationCalls(), 
+				-1E-10, 
+				new CountCrossoverCalls(),
+				1E-10,
 				new TestInitializer(), 
 				new TestFitnessInteger(), 
 				new TestSelectionOp(), 
@@ -811,13 +863,12 @@ public class GenerationalEATests {
 		CountCrossoverCalls crossover = new CountCrossoverCalls();
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.25,
 			crossover,
 			0.25,
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -848,7 +899,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		int oldMutationCount = mutation.count;
@@ -917,13 +968,12 @@ public class GenerationalEATests {
 		// Deliberately odd compared to other tests.
 		final int N = 101;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.5, 
 			crossover,
 			0.5,
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -952,7 +1002,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);
@@ -1013,13 +1063,12 @@ public class GenerationalEATests {
 		// Deliberately odd compared to other tests.
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.0, 
 			crossover,
 			1.0,
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -1049,7 +1098,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);
@@ -1112,13 +1161,12 @@ public class GenerationalEATests {
 		// Deliberately odd compared to other tests.
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			Math.ulp(1.0), 
 			crossover,
 			1.0-Math.ulp(1.0),
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -1148,7 +1196,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);
@@ -1197,308 +1245,6 @@ public class GenerationalEATests {
 		
 		assertEquals(3, selection.initCount);
 	}
-	
-	// double fitness, full generation but explicitly passing false for mutually exclusive ops
-	
-	@Test
-	public void testBothOpsDoubleFitness_MutualFalse() {
-		TestObject.reinit();
-		ProgressTracker<TestObject> tracker = new ProgressTracker<TestObject>();
-		TestSelectionOp selection = new TestSelectionOp();
-		TestFitnessDouble f = new TestFitnessDouble();
-		TestInitializer initializer = new TestInitializer();
-		CountMutationCalls mutation = new CountMutationCalls();
-		CountCrossoverCalls crossover = new CountCrossoverCalls();
-		final int N = 100;
-		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
-			N, 
-			mutation, 
-			0.5,
-			crossover,
-			0.5,
-			false,
-			initializer, 
-			f, 
-			selection, 
-			tracker
-		);
-		
-		assertTrue(tracker == ea.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		ea.setProgressTracker(tracker);
-		assertTrue(tracker == ea.getProgressTracker());
-		assertTrue(f.getProblem() == ea.getProblem());
-		
-		assertEquals(0L, ea.getTotalRunLength());
-		
-		assertEquals(0, selection.initCount);
-		
-		// optimize
-		SolutionCostPair<TestObject> solution = ea.optimize(3);
-		assertEquals(N, initializer.count);
-		assertTrue(mutation.count > 0 && mutation.count < 3*N);
-		assertTrue(crossover.count > 0 && crossover.count < 3*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(3, selection.calledCount);
-		assertEquals(tracker.getSolutionCostPair().getSolution(), solution.getSolution());
-		assertTrue(tracker.getSolutionCostPair().getSolution() != solution.getSolution());
-		assertEquals(1, selection.initCount);
-		
-		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
-		
-		// reoptimize
-		int oldMutationCount = mutation.count;
-		int oldCrossoverCount = crossover.count;
-		solution = ea.reoptimize(5);
-		assertNotNull(solution);
-		assertEquals(N, initializer.count);
-		assertTrue(mutation.count > oldMutationCount && mutation.count < 8*N);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 8*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(8, selection.calledCount);
-		assertEquals(2, selection.initCount);
-		
-		// another optimize
-		oldMutationCount = mutation.count;
-		oldCrossoverCount = crossover.count;
-		solution = ea.optimize(2);
-		assertNotNull(solution);
-		assertEquals(2*N, initializer.count);
-		assertTrue(mutation.count > oldMutationCount && mutation.count < 10*N);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 10*N);
-		assertEquals(f.count, 2*N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(10, selection.calledCount);
-		assertEquals(3, selection.initCount);
-		
-		tracker.stop();
-		assertNull(ea.optimize(2));
-		assertNull(ea.reoptimize(2));
-		assertEquals(3, selection.initCount);
-		
-		// Try the split version
-		assertEquals(0L, ea2.getTotalRunLength());
-		assertNull(ea2.optimize(2));
-		assertNull(ea2.reoptimize(2));
-		tracker.start();
-		oldMutationCount = mutation.count;
-		oldCrossoverCount = crossover.count;
-		int oldFCount = f.count;
-		solution = ea2.optimize(1);
-		assertEquals((long)(f.count-oldFCount), ea2.getTotalRunLength());
-		assertNotNull(solution);
-		// These should change since it should have been split.
-		assertEquals(2*N, initializer.count);
-		assertEquals(oldMutationCount, mutation.count);
-		assertEquals(oldCrossoverCount, crossover.count);
-		// This can be shared so should increase
-		assertTrue(f.count > oldFCount);
-		
-		assertEquals(3, selection.initCount);
-	}
-	
-	@Test
-	public void testBothOpsAlwaysMutateDoubleFitness_MutualFalse() {
-		TestObject.reinit();
-		ProgressTracker<TestObject> tracker = new ProgressTracker<TestObject>();
-		TestSelectionOp selection = new TestSelectionOp();
-		TestFitnessDouble f = new TestFitnessDouble();
-		TestInitializer initializer = new TestInitializer();
-		CountMutationCalls mutation = new CountMutationCalls();
-		CountCrossoverCalls crossover = new CountCrossoverCalls();
-		final int N = 100;
-		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
-			N, 
-			mutation, 
-			1.0, 
-			crossover,
-			0.5,
-			false,
-			initializer, 
-			f, 
-			selection, 
-			tracker
-		);
-		
-		assertTrue(tracker == ea.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		ea.setProgressTracker(tracker);
-		assertTrue(tracker == ea.getProgressTracker());
-		assertTrue(f.getProblem() == ea.getProblem());
-		
-		assertEquals(0L, ea.getTotalRunLength());
-		
-		assertEquals(0, selection.initCount);
-		
-		// optimize
-		SolutionCostPair<TestObject> solution = ea.optimize(3);
-		assertEquals(N, initializer.count);
-		assertEquals(3*N, mutation.count);
-		assertTrue(crossover.count > 0 && crossover.count < 3*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(3, selection.calledCount);
-		assertEquals(tracker.getSolutionCostPair().getSolution(), solution.getSolution());
-		assertTrue(tracker.getSolutionCostPair().getSolution() != solution.getSolution());
-		assertEquals(1, selection.initCount);
-		
-		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
-		
-		// reoptimize
-		int oldCrossoverCount = crossover.count;
-		solution = ea.reoptimize(5);
-		assertNotNull(solution);
-		assertEquals(N, initializer.count);
-		assertEquals(8*N, mutation.count);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 8*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(8, selection.calledCount);
-		assertEquals(2, selection.initCount);
-		
-		// another optimize
-		oldCrossoverCount = crossover.count;
-		solution = ea.optimize(2);
-		assertNotNull(solution);
-		assertEquals(2*N, initializer.count);
-		assertEquals(10*N, mutation.count);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 10*N);
-		assertEquals(f.count, 2*N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(10, selection.calledCount);
-		assertEquals(3, selection.initCount);
-		
-		tracker.stop();
-		assertNull(ea.optimize(2));
-		assertNull(ea.reoptimize(2));
-		assertEquals(3, selection.initCount);
-		
-		// Try the split version
-		assertEquals(0L, ea2.getTotalRunLength());
-		assertNull(ea2.optimize(2));
-		assertNull(ea2.reoptimize(2));
-		tracker.start();
-		int oldMutationCount = mutation.count;
-		oldCrossoverCount = crossover.count;
-		int oldFCount = f.count;
-		solution = ea2.optimize(1);
-		assertEquals((long)(f.count-oldFCount), ea2.getTotalRunLength());
-		assertNotNull(solution);
-		// These should change since it should have been split.
-		assertEquals(2*N, initializer.count);
-		assertEquals(oldMutationCount, mutation.count);
-		assertEquals(oldCrossoverCount, crossover.count);
-		// This can be shared so should increase
-		assertTrue(f.count > oldFCount);
-		
-		assertEquals(3, selection.initCount);
-	}
-	
-	@Test
-	public void testBothOps_1_0_DoubleFitness_MutualFalse() {
-		TestObject.reinit();
-		ProgressTracker<TestObject> tracker = new ProgressTracker<TestObject>();
-		TestSelectionOp selection = new TestSelectionOp();
-		TestFitnessDouble f = new TestFitnessDouble();
-		TestInitializer initializer = new TestInitializer();
-		CountMutationCalls mutation = new CountMutationCalls();
-		CountCrossoverCalls crossover = new CountCrossoverCalls();
-		
-		// Deliberately odd compared to other tests.
-		final int N = 101;
-		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
-			N, 
-			mutation, 
-			1.0, 
-			crossover,
-			1.0,
-			false,
-			initializer, 
-			f, 
-			selection, 
-			tracker
-		);
-		
-		assertTrue(tracker == ea.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		ea.setProgressTracker(tracker);
-		assertTrue(tracker == ea.getProgressTracker());
-		assertTrue(f.getProblem() == ea.getProblem());
-		
-		assertEquals(0L, ea.getTotalRunLength());
-		
-		assertEquals(0, selection.initCount);
-		
-		// optimize
-		SolutionCostPair<TestObject> solution = ea.optimize(3);
-		assertEquals(N, initializer.count);
-		assertEquals(3*N, mutation.count);
-		assertEquals(3*N-3, crossover.count);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(3, selection.calledCount);
-		assertEquals(tracker.getSolutionCostPair().getSolution(), solution.getSolution());
-		assertTrue(tracker.getSolutionCostPair().getSolution() != solution.getSolution());
-		assertEquals(1, selection.initCount);
-		
-		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
-		
-		// reoptimize
-		solution = ea.reoptimize(5);
-		assertNotNull(solution);
-		assertEquals(N, initializer.count);
-		assertEquals(8*N, mutation.count);
-		assertEquals(8*N-8, crossover.count);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(8, selection.calledCount);
-		assertEquals(2, selection.initCount);
-		
-		// another optimize
-		solution = ea.optimize(2);
-		assertNotNull(solution);
-		assertEquals(2*N, initializer.count);
-		assertEquals(10*N, mutation.count);
-		assertEquals(10*N-10, crossover.count);
-		assertEquals(f.count, 2*N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(10, selection.calledCount);
-		assertEquals(3, selection.initCount);
-		
-		tracker.stop();
-		assertNull(ea.optimize(2));
-		assertNull(ea.reoptimize(2));
-		assertEquals(3, selection.initCount);
-		
-		// Try the split version
-		assertEquals(0L, ea2.getTotalRunLength());
-		assertNull(ea2.optimize(2));
-		assertNull(ea2.reoptimize(2));
-		tracker.start();
-		int oldMutationCount = mutation.count;
-		int oldCrossoverCount = crossover.count;
-		int oldFCount = f.count;
-		solution = ea2.optimize(1);
-		assertEquals((long)(f.count-oldFCount), ea2.getTotalRunLength());
-		assertNotNull(solution);
-		// These should change since it should have been split.
-		assertEquals(2*N, initializer.count);
-		assertEquals(oldMutationCount, mutation.count);
-		assertEquals(oldCrossoverCount, crossover.count);
-		// This can be shared so should increase
-		assertTrue(f.count > oldFCount);
-		
-		assertEquals(3, selection.initCount);
-	}
-	
 	
 	// int fitness
 	
@@ -1988,13 +1734,12 @@ public class GenerationalEATests {
 		CountCrossoverCalls crossover = new CountCrossoverCalls();
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.25,
 			crossover,
 			0.25,
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -2025,7 +1770,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		int oldMutationCount = mutation.count;
@@ -2094,13 +1839,12 @@ public class GenerationalEATests {
 		// Deliberately odd compared to other tests.
 		final int N = 101;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.5, 
 			crossover,
 			0.5,
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -2129,7 +1873,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);
@@ -2190,13 +1934,12 @@ public class GenerationalEATests {
 		// Deliberately odd compared to other tests.
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.0, 
 			crossover,
 			1.0,
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -2226,7 +1969,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);
@@ -2289,13 +2032,12 @@ public class GenerationalEATests {
 		// Deliberately odd compared to other tests.
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea = new GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			Math.ulp(1.0), 
 			crossover,
 			1.0-Math.ulp(1.0),
-			true,
 			initializer, 
 			f, 
 			selection, 
@@ -2325,7 +2067,7 @@ public class GenerationalEATests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalNANDOperatorsEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);
@@ -2374,308 +2116,6 @@ public class GenerationalEATests {
 		
 		assertEquals(3, selection.initCount);
 	}
-	
-	// int fitness, full generation but explicitly passing false for mutually exclusive ops
-	
-	@Test
-	public void testBothOpsIntegerFitness_MutualFalse() {
-		TestObject.reinit();
-		ProgressTracker<TestObject> tracker = new ProgressTracker<TestObject>();
-		TestSelectionOp selection = new TestSelectionOp();
-		TestFitnessInteger f = new TestFitnessInteger();
-		TestInitializer initializer = new TestInitializer();
-		CountMutationCalls mutation = new CountMutationCalls();
-		CountCrossoverCalls crossover = new CountCrossoverCalls();
-		final int N = 100;
-		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
-			N, 
-			mutation, 
-			0.5,
-			crossover,
-			0.5,
-			false,
-			initializer, 
-			f, 
-			selection, 
-			tracker
-		);
-		
-		assertTrue(tracker == ea.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		ea.setProgressTracker(tracker);
-		assertTrue(tracker == ea.getProgressTracker());
-		assertTrue(f.getProblem() == ea.getProblem());
-		
-		assertEquals(0L, ea.getTotalRunLength());
-		
-		assertEquals(0, selection.initCount);
-		
-		// optimize
-		SolutionCostPair<TestObject> solution = ea.optimize(3);
-		assertEquals(N, initializer.count);
-		assertTrue(mutation.count > 0 && mutation.count < 3*N);
-		assertTrue(crossover.count > 0 && crossover.count < 3*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(3, selection.calledCount);
-		assertEquals(tracker.getSolutionCostPair().getSolution(), solution.getSolution());
-		assertTrue(tracker.getSolutionCostPair().getSolution() != solution.getSolution());
-		assertEquals(1, selection.initCount);
-		
-		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
-		
-		// reoptimize
-		int oldMutationCount = mutation.count;
-		int oldCrossoverCount = crossover.count;
-		solution = ea.reoptimize(5);
-		assertNotNull(solution);
-		assertEquals(N, initializer.count);
-		assertTrue(mutation.count > oldMutationCount && mutation.count < 8*N);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 8*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(8, selection.calledCount);
-		assertEquals(2, selection.initCount);
-		
-		// another optimize
-		oldMutationCount = mutation.count;
-		oldCrossoverCount = crossover.count;
-		solution = ea.optimize(2);
-		assertNotNull(solution);
-		assertEquals(2*N, initializer.count);
-		assertTrue(mutation.count > oldMutationCount && mutation.count < 10*N);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 10*N);
-		assertEquals(f.count, 2*N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(10, selection.calledCount);
-		assertEquals(3, selection.initCount);
-		
-		tracker.stop();
-		assertNull(ea.optimize(2));
-		assertNull(ea.reoptimize(2));
-		assertEquals(3, selection.initCount);
-		
-		// Try the split version
-		assertEquals(0L, ea2.getTotalRunLength());
-		assertNull(ea2.optimize(2));
-		assertNull(ea2.reoptimize(2));
-		tracker.start();
-		oldMutationCount = mutation.count;
-		oldCrossoverCount = crossover.count;
-		int oldFCount = f.count;
-		solution = ea2.optimize(1);
-		assertEquals((long)(f.count-oldFCount), ea2.getTotalRunLength());
-		assertNotNull(solution);
-		// These should change since it should have been split.
-		assertEquals(2*N, initializer.count);
-		assertEquals(oldMutationCount, mutation.count);
-		assertEquals(oldCrossoverCount, crossover.count);
-		// This can be shared so should increase
-		assertTrue(f.count > oldFCount);
-		
-		assertEquals(3, selection.initCount);
-	}
-	
-	@Test
-	public void testBothOpsAlwaysMutateIntegerFitness_MutualFalse() {
-		TestObject.reinit();
-		ProgressTracker<TestObject> tracker = new ProgressTracker<TestObject>();
-		TestSelectionOp selection = new TestSelectionOp();
-		TestFitnessInteger f = new TestFitnessInteger();
-		TestInitializer initializer = new TestInitializer();
-		CountMutationCalls mutation = new CountMutationCalls();
-		CountCrossoverCalls crossover = new CountCrossoverCalls();
-		final int N = 100;
-		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
-			N, 
-			mutation, 
-			1.0, 
-			crossover,
-			0.5,
-			false,
-			initializer, 
-			f, 
-			selection, 
-			tracker
-		);
-		
-		assertTrue(tracker == ea.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		ea.setProgressTracker(tracker);
-		assertTrue(tracker == ea.getProgressTracker());
-		assertTrue(f.getProblem() == ea.getProblem());
-		
-		assertEquals(0L, ea.getTotalRunLength());
-		
-		assertEquals(0, selection.initCount);
-		
-		// optimize
-		SolutionCostPair<TestObject> solution = ea.optimize(3);
-		assertEquals(N, initializer.count);
-		assertEquals(3*N, mutation.count);
-		assertTrue(crossover.count > 0 && crossover.count < 3*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(3, selection.calledCount);
-		assertEquals(tracker.getSolutionCostPair().getSolution(), solution.getSolution());
-		assertTrue(tracker.getSolutionCostPair().getSolution() != solution.getSolution());
-		assertEquals(1, selection.initCount);
-		
-		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
-		
-		// reoptimize
-		int oldCrossoverCount = crossover.count;
-		solution = ea.reoptimize(5);
-		assertNotNull(solution);
-		assertEquals(N, initializer.count);
-		assertEquals(8*N, mutation.count);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 8*N);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(8, selection.calledCount);
-		assertEquals(2, selection.initCount);
-		
-		// another optimize
-		oldCrossoverCount = crossover.count;
-		solution = ea.optimize(2);
-		assertNotNull(solution);
-		assertEquals(2*N, initializer.count);
-		assertEquals(10*N, mutation.count);
-		assertTrue(crossover.count > oldCrossoverCount && crossover.count < 10*N);
-		assertEquals(f.count, 2*N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(10, selection.calledCount);
-		assertEquals(3, selection.initCount);
-		
-		tracker.stop();
-		assertNull(ea.optimize(2));
-		assertNull(ea.reoptimize(2));
-		assertEquals(3, selection.initCount);
-		
-		// Try the split version
-		assertEquals(0L, ea2.getTotalRunLength());
-		assertNull(ea2.optimize(2));
-		assertNull(ea2.reoptimize(2));
-		tracker.start();
-		int oldMutationCount = mutation.count;
-		oldCrossoverCount = crossover.count;
-		int oldFCount = f.count;
-		solution = ea2.optimize(1);
-		assertEquals((long)(f.count-oldFCount), ea2.getTotalRunLength());
-		assertNotNull(solution);
-		// These should change since it should have been split.
-		assertEquals(2*N, initializer.count);
-		assertEquals(oldMutationCount, mutation.count);
-		assertEquals(oldCrossoverCount, crossover.count);
-		// This can be shared so should increase
-		assertTrue(f.count > oldFCount);
-		
-		assertEquals(3, selection.initCount);
-	}
-	
-	@Test
-	public void testBothOps_1_0_IntegerFitness_MutualFalse() {
-		TestObject.reinit();
-		ProgressTracker<TestObject> tracker = new ProgressTracker<TestObject>();
-		TestSelectionOp selection = new TestSelectionOp();
-		TestFitnessInteger f = new TestFitnessInteger();
-		TestInitializer initializer = new TestInitializer();
-		CountMutationCalls mutation = new CountMutationCalls();
-		CountCrossoverCalls crossover = new CountCrossoverCalls();
-		
-		// Deliberately odd compared to other tests.
-		final int N = 101;
-		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
-			N, 
-			mutation, 
-			1.0, 
-			crossover,
-			1.0,
-			false,
-			initializer, 
-			f, 
-			selection, 
-			tracker
-		);
-		
-		assertTrue(tracker == ea.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		ea.setProgressTracker(tracker);
-		assertTrue(tracker == ea.getProgressTracker());
-		assertTrue(f.getProblem() == ea.getProblem());
-		
-		assertEquals(0L, ea.getTotalRunLength());
-		
-		assertEquals(0, selection.initCount);
-		
-		// optimize
-		SolutionCostPair<TestObject> solution = ea.optimize(3);
-		assertEquals(N, initializer.count);
-		assertEquals(3*N, mutation.count);
-		assertEquals(3*N-3, crossover.count);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(3, selection.calledCount);
-		assertEquals(tracker.getSolutionCostPair().getSolution(), solution.getSolution());
-		assertTrue(tracker.getSolutionCostPair().getSolution() != solution.getSolution());
-		assertEquals(1, selection.initCount);
-		
-		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
-		
-		// reoptimize
-		solution = ea.reoptimize(5);
-		assertNotNull(solution);
-		assertEquals(N, initializer.count);
-		assertEquals(8*N, mutation.count);
-		assertEquals(8*N-8, crossover.count);
-		assertEquals(f.count, N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(8, selection.calledCount);
-		assertEquals(2, selection.initCount);
-		
-		// another optimize
-		solution = ea.optimize(2);
-		assertNotNull(solution);
-		assertEquals(2*N, initializer.count);
-		assertEquals(10*N, mutation.count);
-		assertEquals(10*N-10, crossover.count);
-		assertEquals(f.count, 2*N + mutation.count + crossover.count);
-		assertEquals((long)f.count, ea.getTotalRunLength());
-		assertEquals(10, selection.calledCount);
-		assertEquals(3, selection.initCount);
-		
-		tracker.stop();
-		assertNull(ea.optimize(2));
-		assertNull(ea.reoptimize(2));
-		assertEquals(3, selection.initCount);
-		
-		// Try the split version
-		assertEquals(0L, ea2.getTotalRunLength());
-		assertNull(ea2.optimize(2));
-		assertNull(ea2.reoptimize(2));
-		tracker.start();
-		int oldMutationCount = mutation.count;
-		int oldCrossoverCount = crossover.count;
-		int oldFCount = f.count;
-		solution = ea2.optimize(1);
-		assertEquals((long)(f.count-oldFCount), ea2.getTotalRunLength());
-		assertNotNull(solution);
-		// These should change since it should have been split.
-		assertEquals(2*N, initializer.count);
-		assertEquals(oldMutationCount, mutation.count);
-		assertEquals(oldCrossoverCount, crossover.count);
-		// This can be shared so should increase
-		assertTrue(f.count > oldFCount);
-		
-		assertEquals(3, selection.initCount);
-	}
-	
 	
 	// HELPER CLASSES